Move 'commands' modules into '__init__'
In the command heirarchy, use `__init__.py` for the main command group
for each level in the command tree. This removes the need for a
distinct `commands.py` module which defines the main group, and
therefore reduces LOC and number of files.
